Added option to randomize the name before the upload
I love and use this little app daily but it always bothered me that it
kept the original file name.

Just added an option, unchecked by default, to randomize the name of
the file before uploading it.

It’s the very first time I open a Swift project, so I took no risk and
kept everything naively simple. Implementation might be a bit rough
though. Also, I’m not sure wether I should commit the Pod files, but
their path is not in the gitignore so I guess I do ?

I added the required method in a separate file because if I come up
with the time to read a bit about the language I’d also like to
implement an Exif & ICC cleaner.