Security question
(Sorry to reach out in this way. I couldn't find out another way to contact the maintainers.)

I just received an update notification that said "This update is important". I just want to ask if this was a legitimate message or if there's something malicious going on?

I'm concerned because I know this project is not actively maintained anymore.

Thank you!